Rename internal headers
